requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-leaflet==0.27.1
djangorestframework==3.12.4
GDAL==3.0.4
greenlet==1.1.2
idna==3.3
Pillow==8.4.0
psycopg2-binary==2.9.1
pyTelegramBotAPI==4.1.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
